[refactor] 멤버 dto 폴더 분리
