+++ src/main/java/ctrlS/totori/member/controller/MemberController.java
@@ -2,10 +2,10 @@ package ctrlS.totori.member.controller;
 
 import ctrlS.totori.global.response.dto.BaseResponse;
 import ctrlS.totori.global.security.CustomUserPrincipal;
-import ctrlS.totori.member.dto.AcornResponse;
-import ctrlS.totori.member.dto.MemberMeResponse;
-import ctrlS.totori.member.dto.UpdateMemberRequest;
-import ctrlS.totori.member.dto.UpdateMemberResponse;
+import ctrlS.totori.member.dto.response.AcornResponse;
+import ctrlS.totori.member.dto.response.MemberMeResponse;
+import ctrlS.totori.member.dto.request.UpdateMemberRequest;
+import ctrlS.totori.member.dto.response.UpdateMemberResponse;
 import ctrlS.totori.member.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
