+++ src/main/java/ctrlS/totori/member/dto/response/UpdateMemberResponse.java
@@ -1,4 +1,4 @@
-package ctrlS.totori.member.dto;
+package ctrlS.totori.member.dto.response;
 
 import ctrlS.totori.member.entity.Member;
