+++ src/main/java/ctrlS/totori/member/service/MemberService.java
@@ -4,10 +4,10 @@ import ctrlS.totori.auth.service.AuthRedisService;
 import ctrlS.totori.global.exception.CustomException;
 import ctrlS.totori.global.exception.ErrorCode;
 import ctrlS.totori.global.security.JwtTokenProvider;
-import ctrlS.totori.member.dto.AcornResponse;
-import ctrlS.totori.member.dto.MemberMeResponse;
-import ctrlS.totori.member.dto.UpdateMemberRequest;
-import ctrlS.totori.member.dto.UpdateMemberResponse;
+import ctrlS.totori.member.dto.response.AcornResponse;
+import ctrlS.totori.member.dto.response.MemberMeResponse;
+import ctrlS.totori.member.dto.request.UpdateMemberRequest;
+import ctrlS.totori.member.dto.response.UpdateMemberResponse;
 import ctrlS.totori.member.entity.Member;
 import ctrlS.totori.member.entity.Role;
 import ctrlS.totori.member.repository.MemberRepository;
